replaced diff_logic with pygit2 implementation
